reorganize code of vectors app
